문 크기 조정

+++ return S_OK TeamPortpolio/PlayerManager.h
@@ -24,8 +24,8 @@ public:
 	float GetPlayerHitRectX() { return m_Player->GetPlayerHitRectX(); }
 	float GetPlayerHitRectY() { return m_Player->GetPlayerHitRectY(); }
 
-	void SetPlayerRectX(int num) { m_Player->SetPlayerRectX(num); }
-	void SetPlayerRectY(int num) { m_Player->SetPlayerRectY(num); }
+	void SetPlayerRectX(float num) { m_Player->SetPlayerRectX(num); }
+	void SetPlayerRectY(float num) { m_Player->SetPlayerRectY(num); }
 
 	int GetPlayerOffensePower() { return m_Player->GetPlayerOffensePower(); }
 	void SetPlayerOffensePower(int addPower) { m_Player->SetPlayerOffensePower(addPower); }
